Adding Variables And Functions BlueprintPure Vs BlueprintCallable
BlueprintPure is generally used as a getter function/where constantly values are changing/fast calculations/They do not modify the members of the class, etc. They do not have execution pins.

BlueprintCallable is generally used when members of the class is to be modified. They have input and output execution pins.